+++ .ci/run.sh
@@ -10,28 +10,41 @@ PYTHON=${PYTHON:-"python3.7"}
 
 TEMP_DIRS=()
 
+function run_plugin_examples() {
+  if [ ! -d "${SRC_ROOT}/${PLUGIN}/examples" ]; then
+    return
+  fi
+  cd "${SRC_ROOT}/${PLUGIN}/examples"
+  if [ -f "requirements.txt" ]; then
+    "${PYTHON}" -m pip install -r requirements.txt
+  fi
+  "${PYTHON}" -m unittest discover -v
+  cd "${SRC_ROOT}/${PLUGIN}"
+}
+
 function run_plugin() {
-  # Create a virtualenv
-  venv_dir="$(mktemp -d)"
-  TEMP_DIRS+=("${venv_dir}")
-  "${PYTHON}" -m venv "${venv_dir}"
-  source "${venv_dir}/bin/activate"
   "${PYTHON}" -m pip install -U pip twine
 
-  "${PYTHON}" -m pip install -U "${SRC_ROOT}"
+  # Install main package
+  "${PYTHON}" -m pip install -U -e "${SRC_ROOT}[dev]"
 
   if [ "x${PLUGIN}" = "xmodel/tensorflow_hub" ]; then
-    "${PYTHON}" -m pip install -U "${SRC_ROOT}/model/tensorflow"
+    "${PYTHON}" -m pip install -U -e "${SRC_ROOT}/model/tensorflow"
   fi
 
-  cd "${PLUGIN}"
-  PACKAGE_NAME=$(dffml service dev setuppy kwarg name setup.py)
-  "${PYTHON}" -m pip install -e .
-  "${PYTHON}" setup.py test
-  "${PYTHON}" -m pip uninstall -y "${PACKAGE_NAME}"
-  cd -
+  cd "${SRC_ROOT}/${PLUGIN}"
 
-  if [ "x${PLUGIN}" = "x." ]; then
+  # Install plugin
+  "${PYTHON}" -m pip install -U -e .
+  # Run the tests
+  "${PYTHON}" setup.py test
+
+  if [ "x${PLUGIN}" != "x." ]; then
+    # Run examples if they exist and we aren't at the root
+    run_plugin_examples
+  else
+    # If we are at the root. Install plugsin and run various integration tests
+
     # Try running create command
     plugin_creation_dir="$(mktemp -d)"
     TEMP_DIRS+=("${plugin_creation_dir}")
@@ -52,22 +65,10 @@ function run_plugin() {
     done
 
     # Install all the plugins so examples can use them
-    "${PYTHON}" -m pip install -U -e "${SRC_ROOT}"
     "${PYTHON}" -m dffml service dev install
 
     # Run the examples
-    cd "${SRC_ROOT}/examples"
-    "${PYTHON}" -m pip install -r requirements.txt
-    "${PYTHON}" -m unittest discover
-    cd "${SRC_ROOT}"
-
-    # Deactivate venv
-    deactivate
-
-    # Create the docs
-    "${PYTHON}" -m pip install -U -e "${SRC_ROOT}[dev]"
-    "${PYTHON}" -m dffml service dev install -user
-    "${SRC_ROOT}/scripts/docs.sh"
+    run_plugin_examples
 
     # Log skipped tests to file
     check_skips="$(mktemp)"
@@ -77,9 +78,21 @@ function run_plugin() {
     "${PYTHON}" -m coverage run setup.py test 2>&1 | tee "${check_skips}"
     "${PYTHON}" -m coverage report -m
 
-    # Fail if any tests were skipped
-    grep -v -q -E '(skipped=.*)' "${check_skips}"
+    # Fail if any tests were skipped or errored
+    skipped=$(grep -E '(skipped=.*)' "${check_skips}" | wc -l)
+    if [ "$skipped" -ne 0 ]; then
+      echo "Tests were skipped" >&2
+      exit 1
+    fi
+
+    errors=$(grep -E '(errors=.*)' "${check_skips}" | wc -l)
+    if [ "$errors" -ne 0 ]; then
+      echo "Tests errored" >&2
+      exit 1
+    fi
   fi
+
+  cd "${SRC_ROOT}"
 
   if [ "x${GITHUB_ACTIONS}" == "xtrue" ] && [ "x${GITHUB_REF}" == "xrefs/heads/master" ]; then
     git status
@@ -118,6 +131,11 @@ function run_whitespace() {
 
 function run_style() {
   black --check "${SRC_ROOT}"
+
+  for filename in $(git ls-files \*.js); do
+    echo "Checking JavaScript file \'${filename}\'"
+    diff <(js-beautify -n -s 2 "${filename}") "${filename}"
+  done
 }
 
 function run_docs() {
@@ -126,6 +144,19 @@ function run_docs() {
   cd "${SRC_ROOT}"
   "${PYTHON}" -m pip install --prefix=~/.local -U -e "${SRC_ROOT}[dev]"
   "${PYTHON}" -m dffml service dev install -user
+
+  last_release=$("${PYTHON}" -m dffml service dev setuppy kwarg version setup.py)
+
+  # Doctests
+  ./scripts/doctest.sh
+
+  # Fail if there are any changes to the Git repo
+  changes=$(git status --porcelain | wc -l)
+  if [ "$changes" -ne 0 ]; then
+    echo "Running docs.py resulted in changes to the Git repo" >&2
+    echo "Need to run ./scripts/docs.sh and commit changes" >&2
+    exit 1
+  fi
 
   # Make master docs
   master_docs="$(mktemp -d)"
@@ -139,9 +170,13 @@ function run_docs() {
   TEMP_DIRS+=("${release_docs}")
   rm -rf pages
   git clean -fdx
-  git checkout $(git describe --abbrev=0 --tags --match '*.*.*')
+  git reset --hard HEAD
+  echo "Checking out last release ${last_release}"
+  git checkout "${last_release}"
   git clean -fdx
   git reset --hard HEAD
+  # Uninstall dffml
+  "${PYTHON}" -m pip uninstall -y dffml
   # Remove .local to force install of correct dependency versions
   rm -rf ~/.local
   "${PYTHON}" -m pip install --prefix=~/.local -U -e "${SRC_ROOT}[dev]"
@@ -182,12 +217,16 @@ function run_docs() {
   export GIT_SSH_COMMAND="${GIT_SSH_COMMAND} -o IdentityFile=~/.ssh/github_dffml"
 
   git remote set-url origin git@github.com:intel/dffml
-  git push
+  git push -f
 
   cd -
 
   git reset --hard HEAD
   git checkout master
+}
+
+function run_lines() {
+  "${PYTHON}" ./scripts/check_literalincludes.py
 }
 
 function cleanup_temp_dirs() {
@@ -212,4 +251,6 @@ elif [ "x${STYLE}" != "x" ]; then
   run_style
 elif [ "x${DOCS}" != "x" ]; then
   run_docs
+elif [ "x${LINES}" != "x" ]; then
+  run_lines
 fi