+++ dffml/df/base.py
@@ -12,6 +12,7 @@ from typing import (
     Optional,
     Set,
 )
+from contextlib import asynccontextmanager
 
 from .exceptions import NotOpImp
 from .types import Operation, Input, Parameter, Stage, Definition
@@ -85,6 +86,19 @@ class OperationImplementationContext(BaseDataFlowObjectContext):
         with keys matching the input and output parameters of the Operation
         object associated with this operation implementation context.
         """
+
+    @asynccontextmanager
+    async def subflow(self, dataflow):
+        """
+        Registers subflow `dataflow` with parent flow and yields an instance of `BaseOrchestratorContext`
+
+        >>> async def my_operation(arg):
+        ...     async with self.subflow(self.config.dataflow) as octx:
+        ...         return octx.run({"ctx_str": []})
+        """
+        async with self.octx.parent(dataflow) as octx:
+            self.octx.subflows[self.parent.op.instance_name] = octx
+            yield octx
 
 
 class FailedToLoadOperationImplementation(Exception):
@@ -298,11 +312,6 @@ def isopimp(item):
     """
     Similar to inspect.isclass and that family of functions. Returns true if
     item is a subclass of OperationImpelmentation.
-
-    >>> # Get all operation implementations imported in a file
-    >>> list(map(lambda item: item[1],
-    >>>          inspect.getmembers(sys.modules[__name__],
-    >>>                             predicate=isopimp)))
     """
     return bool(
         (
@@ -322,11 +331,6 @@ def isoperation(item):
     """
     Similar to inspect.isclass and that family of functions. Returns true if
     item is an instance of Operation.
-
-    >>> # Get all operations imported in a file
-    >>> list(map(lambda item: item[1],
-    >>>          inspect.getmembers(sys.modules[__name__],
-    >>>                             predicate=isoperation)))
     """
     return bool(isinstance(item, Operation) and item is not Operation)
 
@@ -335,11 +339,6 @@ def isopwraped(item):
     """
     Similar to inspect.isclass and that family of functions. Returns true if a
     function has been wrapped with `op`.
-
-    >>> # Get all functions imported in a file that have been wrapped with `op`
-    >>> list(map(lambda item: item[1],
-    >>>          inspect.getmembers(sys.modules[__name__],
-    >>>                             predicate=isopwraped)))
     """
     return bool(
         getattr(item, "op", False)
@@ -789,4 +788,13 @@ class BaseOrchestratorContext(BaseDataFlowObjectContext):
 
 @base_entry_point("dffml.orchestrator", "orchestrator")
 class BaseOrchestrator(BaseDataFlowObject):
-    pass  # pragma: no cov
+    @classmethod
+    async def run(cls, dataflow, inputs, *, config=None, **kwargs):
+        if config is None:
+            self = cls.withconfig({})
+        else:
+            self = cls(config=config, **kwargs)
+        async with self as orchestrator:
+            async with orchestrator(dataflow) as octx:
+                async for ctx, results in octx.run(inputs):
+                    yield ctx, results